backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
werkzeug==3.0.1
email-validator==2.1.0
python-dateutil==2.8.2
